Add keymap for `winry/winry315`
Add a custom keymap for the Winry315 macropad.  This keymap provides the
following layers:

  - Numeric keypad layer (as fits into the 5×3 layout — the “0” and “.”
    keys are at the sides of the 3×3 block with “1”…“9” instead of on
    the row below; the keys for math operations are not present at all).

  - Number pad layer with the same layout as above, but using the number
    row keycodes instead of numpad (so that it works irrespective of the
    Num Lock state).

  - Function key layer (F1…F12 and PrtSc).

  - Extended function key layer (F13…F24 and PrtSc).

  - Three empty layers which can be filled in using VIA.

  - The “adjust” layer which can be used to select one of the above
    layers, and also has the Num Lock key and math operation keys.

There is also some code to enable RGB matrix indicators on the adjust
layer, so that the state of Num Lock and the selected layer could be
determined by holding the layer switch key.

VIA is enabled, so that the keymap could be changed without recompiling
the firmware.